Run a deploy script on push to main
After main is updated, connect to the hasworn server and run the deploy
script.

The hasworn code is checked out in the root of the homedir of the
hasworn user, hence gitignore containing bash files. This is so as I
make changes to the "live" service as I am figuring out what needs to
happen, I can run `git status`. The hasworn user is locked down so it
can *only* run the script on inbound ssh requests, and that file
cannot be changed by the hasworn user (I have to connect to the server
and `su - hasworn` to become it).

Django is put behind nginx, and in production compose uses gunicorn
not the runserver. The app is restarted in what should be a zero
downtime manner by:

  * using compose to scale it to two serving containers, which yields
    the newer version, and waiting until it is serving the homepage
    200 OK (if it never does, the deploy fails and the old container
    continues serving the app)
  * reloading nginx so it is using both containers to serve the app
  * stopping the old container, and scaling back down to one
  * reloading nginx again, so it doesn't try to use the now-stopped
    container