Add users and basic login/logout
The default django user model believes names consist of First & Last.
This is incorrect, so the app uses a custom user model that treats
name as a simple text field that can be null.

Adding a custom user model also allows the easy addition of extra
attributes in the future, such as being able to build out a
customisable profile.

As the username will form part of the URL to their hasworn profile (ie
the "norm" part of https://norm.hasworn.com) restrict it to ascii
letters, numbers, and hyphens, and it must start with a letter and be
a minimum of two characters.