Separate the restart from the deploy script
A redeploy that changed the deploy script would be in a chicken-and-egg
scenario. It is already running when the code underneath it changes,
including itself. So now the wrong deploy script is running.

Separating out the restart logic from the act of triggering a deploy
should alleviate this scenario almost all the time, except when the
purposefully minimal `deploy.sh` logic might need to change.