Set up "baked" site serving
Generated user profiles will be served statically rather than from the
django app. A named docker volume will be shared between the app and
nginx for this purpose. As a test during getting this working, a file
was created by hand to prove the concept.

Also, in order to be able to use a proper hostname for the app in
development (i.hasworn.test, similar to i.hasworn.com in production),
make use of the nginx container images use of templates to create files
in /etc/nginx/conf.d, substituting "$SERVER_NAME" into the server block
for the app.