Add HTTPS using Let's Encrypt certificates
Redirects any http://i.hasworn.com URL to https://i.hasworn.com.

As this needs a wildcard (*.hasworn.com) certificate, the DNS challenge
has to be used. I found the Lego (https://github.com/go-acme/lego)
client easier to use than certbot, and seemed to require fewer
permissions in the IAM role to make DNS changes (certbot wanted
ListHostedZones, for example).

The IAM policy needed by Lego to make the DNS changes to satisfy a
challenge is: (from https://go-acme.github.io/lego/dns/route53/#policy)

    {
        "Version": "2012-10-17",
        "Statement": [
            {
                "Sid": "",
                "Effect": "Allow",
                "Action": [
                    "route53:ListResourceRecordSets",
                    "route53:GetChange",
                    "route53:ChangeResourceRecordSets"
                ],
                "Resource": [
                    "arn:aws:route53:::hostedzone/*",
                    "arn:aws:route53:::change/*"
                ]
            },
            {
                "Sid": "",
                "Effect": "Allow",
                "Action": "route53:ListHostedZonesByName",
                "Resource": "*"
            }
        ]
    }

This might be able to be further restricted to only the domain in
question under ListHostedZonesByName, but I'm happy to reuse a single
IAM role for all of my Let's Encrypt needs.

Although this commit removes the SERVER_NAME variable and the need to
use the substitution for nginx configuration templates, it is still a
tidy way of including configuration in the nginx image so is kept.