Add the skeleton of the generated site
Each user of hasworn will have a public site generated with the details
of the clothing they have worn and when. This site will be generated
as details are entered, and served statically from nginx.

The idea is that in normal circumstances, someone will update their
data once a day, so there is no need to repeatedly hit the database
when showing what they have worn as nothing will have changed.