Generate a subset when a wearing is added
Once there are a lot of worn items in a user's collection, generating
the entire static site can take many seconds. Instead generate only
those pages materially affected, that of the clothing itself, and the
indexes that show it.

The navigation across the static site will then become slightly out
of date (the "...205 times..." count will be inaccurate on pages
not regenerated) but that can be fixed later.