Add images to Clothing
Clothing should have images, and the images should be easily
replaceable rather than hard-coded uploads from the draft version
of hasworn.

In development, the images are served from the same i.hasworn.test
host as the user and admin interfaces.

In production, the images are uploaded to an S3 bucket, served up as
https://hasworn.m17s.net.