Store the count of days between wearings
When a Wearing is saved, it looks for the previous time the same
Clothing was worn, calculates the days between, and stores that on
the model (so it is available to be used in templates, and in aggregate
database lookups).

Tests that when a retrospective wearing is added it does not break the
count.

(Changing/removing wearings through the admin interface can still break
these counts, but users can't do that, so it is not yet tested.)