Generate wearer sites asynchronously
Introduce celery to take care of asynchronous wearer site rebuilds.
This also adds redis to the mixture, but so far is only storing celery
job information so should consume almost no extra memory beyond that
of the redis instance itself.

Whenever a new wearing is added (or one removed), a rebuild of just
that clothing, the year affected, and the wearer homepage is scheduled
to run ASAP, and a full rebuild of the wearer site in 5 minutes.

The full rebuild includes the value of `wearer.last_update`, and if that
is no longer the case other updates have happened in the time since
scheduling the rebuild and now. This aborts the rebuild, so that a later
queued build will take care of it (this is intended to stop multiple
updates causing multiple rebuilds).