Regenerate the apex last when deploying
The apex page contains the SHA of the current version of hasworn,
and can serve as a quick lookup to see what is deployed. But a deploy
can always fail after it had been generated (eg `restart celery` failed
when there was no celery in the currently running compose arrangement).
So generate it last to better indicate which version is running, and
to serve as an external successful deployment marker.